some basic organization and deduplication
