+++ src/types.ts
@@ -1,0 +1,7 @@
+export type Ticket = {
+  value: number;
+  nonce: number;
+  receiver: string;
+  sender: string;
+  escrowHash: string;
+};
