+++ test/l2.test.ts
@@ -2,6 +2,8 @@ import { ethers } from "hardhat";
 const { expect } = require("chai");
 
 import { Contract } from "ethers";
+import { RECEIVER_PK, SENDER_PK } from "../src/constants";
+import { hashTicket, signData } from "../src/utils";
 
 type Ticket = {
   value: number;
@@ -11,17 +13,8 @@ type Ticket = {
   escrowHash: string;
 };
 
-const abiType = [
-  "tuple(uint256 nonce,uint256 value,address receiver, address sender, bytes32 escrowHash)",
-];
-const receiverWallet = new ethers.Wallet( // Bob
-  "0x91f47a1911c0fd985b34c25962f661f0de606f7ad38ba156902dff48b4d05f97",
-  ethers.provider
-);
-const senderWallet = new ethers.Wallet( // Alice
-  "0xf3d5b8ba24833578a22960b2c7a8be1ebb7907ffe0b346111b8839e981b28b0c",
-  ethers.provider
-);
+const receiverWallet = new ethers.Wallet(RECEIVER_PK, ethers.provider); // Bob
+const senderWallet = new ethers.Wallet(SENDER_PK, ethers.provider); // Alice
 
 const preimage = ethers.utils.hashMessage("Some secret preimage");
 const escrowHash = ethers.utils.keccak256(
@@ -48,17 +41,7 @@ describe("L2 Contract", function () {
       escrowHash: escrowHash,
     };
 
-    const ticketHash = ethers.utils.keccak256(
-      ethers.utils.defaultAbiCoder.encode(abiType, [
-        [
-          ticket.value,
-          ticket.nonce,
-          ticket.receiver,
-          ticket.sender,
-          ticket.escrowHash,
-        ],
-      ])
-    );
+    const ticketHash = hashTicket(ticket);
 
     await l2Contract.lockFundsInEscrow(
       // Alice locks, and escrowHash is written to chain
@@ -91,13 +74,3 @@ describe("L2 Contract", function () {
     ).to.be.revertedWith("Tickets must be distinct");
   });
 });
-
-// TODO: This was stolen from our old nitro protocol repo
-// Probably a cleaner way of signing it
-function signData(hashedData: string, privateKey: string): any {
-  const signingKey = new ethers.utils.SigningKey(privateKey);
-  const hashedMessage = ethers.utils.hashMessage(
-    ethers.utils.arrayify(hashedData)
-  );
-  return ethers.utils.splitSignature(signingKey.signDigest(hashedMessage));
-}
